+++ packages/config/src/chains/details/mainnet.ts
@@ -1,5 +1,5 @@
-import { nativeCurrencyTemplate } from '../../constants/tokens'
-import { ChainInfo, SupportedChainId } from '../types'
+import { nativeCurrencyTemplate } from '../../constants/nativeCurrencyTemplate'
+import { EvmChainInfo, SupportedChainId } from '../types'
 import { RAW_CHAINS_FILES_PATH } from '../../constants/paths'
 
 const ethereumLogo = `${RAW_CHAINS_FILES_PATH}/images/mainnet-logo.svg`
@@ -9,7 +9,7 @@ const ethereumLogo = `${RAW_CHAINS_FILES_PATH}/images/mainnet-logo.svg`
  *
  * See also https://github.com/wevm/viem/blob/main/src/chains/definitions/mainnet.ts
  */
-export const mainnet: ChainInfo = {
+export const mainnet: EvmChainInfo = {
   id: SupportedChainId.MAINNET,
   label: 'Ethereum',
   eip155Label: 'Ethereum Mainnet',