+++ packages/order-book/src/quoteAmountsAndCosts/quoteAmountsAndCosts.types.ts
@@ -1,0 +1,32 @@
+import type { OrderParameters } from '../generated'
+
+export type OrderAmountsBig = { sellAmount: bigint; buyAmount: bigint }
+
+export interface QuoteParameters {
+  sellDecimals: number
+  buyDecimals: number
+  orderParams: OrderParameters
+  protocolFeeBps: number | undefined
+}
+
+export interface QuoteAmountsAndCostsParams {
+  orderParams: OrderParameters
+  protocolFeeBps: number | undefined
+  partnerFeeBps: number | undefined
+  slippagePercentBps: number
+}
+
+export interface QuotePriceParams {
+  numerator: bigint
+  denominator: bigint
+}
+
+export interface QuoteAmountsWithNetworkCosts {
+  isSell: boolean
+  quotePriceParams: QuotePriceParams
+  networkCostAmount: bigint
+  sellAmountBeforeNetworkCosts: bigint
+  buyAmountAfterNetworkCosts: bigint
+  sellAmountAfterNetworkCosts: bigint
+  buyAmountBeforeNetworkCosts: bigint
+}