+++ packages/subgraph/CHANGELOG.md
@@ -1,5 +1,45 @@
 # Changelog
 
+## [0.3.8](https://github.com/cowprotocol/cow-sdk/compare/sdk-subgraph-v0.3.7...sdk-subgraph-v0.3.8) (2026-03-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-common bumped to 0.7.0
+    * @cowprotocol/sdk-config bumped to 0.10.0
+
+## [0.3.7](https://github.com/cowprotocol/cow-sdk/compare/sdk-subgraph-v0.3.6...sdk-subgraph-v0.3.7) (2026-03-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-common bumped to 0.6.3
+    * @cowprotocol/sdk-config bumped to 0.9.0
+
+## [0.3.6](https://github.com/cowprotocol/cow-sdk/compare/sdk-subgraph-v0.3.5...sdk-subgraph-v0.3.6) (2026-02-20)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-common bumped to 0.6.2
+    * @cowprotocol/sdk-config bumped to 0.8.1
+
+## [0.3.5](https://github.com/cowprotocol/cow-sdk/compare/sdk-subgraph-v0.3.4...sdk-subgraph-v0.3.5) (2026-02-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-common bumped to 0.6.1
+    * @cowprotocol/sdk-config bumped to 0.8.0
+
 ## [0.3.4](https://github.com/cowprotocol/cow-sdk/compare/sdk-subgraph-v0.3.3...sdk-subgraph-v0.3.4) (2026-02-05)